+++ CHANGELOG.md
@@ -27,9 +27,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   to create smaller datasets from bigger ones
   (<https://github.com/openvinotoolkit/datumaro/pull/636>,
    <https://github.com/openvinotoolkit/datumaro/pull/640>)
-- Support for downloading the ImageNetV2 and COCO datasets
-  (<https://github.com/openvinotoolkit/datumaro/pull/653>,
-   <https://github.com/openvinotoolkit/datumaro/pull/659>)
+- Corrector: Delete fix transforms (`delete_image`, `delete_annotation`, `delete_attribute`)
+  to delete dataset items/annoations/attributes with annotation errors
+  (<https://github.com/openvinotoolkit/datumaro/pull/636>,
+   <https://github.com/openvinotoolkit/datumaro/pull/640>)
+- Support for downloading the ImageNetV2 dataset
+  (<https://github.com/openvinotoolkit/datumaro/pull/653>)
 
 ### Changed
 - Allowed direct file paths in `datum import`. Such sources are imported like